+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.0.2] - 2025-04-30
+### Added
+- fix apkg upload bug
+- delete unzipped folder after 
+- delete unused media files
+
 ## [1.0.0] - 2025-04-30
 ### Added
 - Initial release